requirements.txt
Django==2.2.3
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.18
django-nine==0.2.2
djangorestframework==3.9.4
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
ipaddress==1.0.22
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
